+++ src/language/es-format.cjs
@@ -1,5 +1,8 @@
+const rule = require('../util/rule.cjs');
+
 /** @type {import('eslint').Linter.RulesRecord} */
 module.exports = {
+	'array-bracket-spacing': ['error'],
 	'arrow-parens': ['error', 'always'],
 	'arrow-spacing': ['error', {
 		before: true,
@@ -8,11 +11,11 @@ module.exports = {
 	'brace-style': ['error', 'stroustrup', {
 		allowSingleLine: false
 	}],
-	'comma-spacing': ['error', {
+	...rule.override('comma-dangle', '@typescript-eslint/comma-dangle', ['error', 'never']),
+	...rule.override('comma-spacing', '@typescript-eslint/comma-spacing', ['error', {
 		before: false,
 		after: true
-	}],
-	'comma-dangle': ['error', 'never'],
+	}]),
 	curly: 'error',
 	'key-spacing': 'error',
 	'lines-between-class-members': ['error', 'always', {
@@ -22,19 +25,17 @@ module.exports = {
 	'no-empty': ['error', {
 		allowEmptyCatch: true
 	}],
+	...rule.override('no-extra-parens', '@typescript-eslint/no-extra-parens', 'error'),
 	'no-useless-concat': 'error',
-	'object-curly-spacing': ['error', 'always'],
+	...rule.override('object-curly-spacing', '@typescript-eslint/object-curly-spacing', ['error', 'always']),
 	'padded-blocks': ['error', {
 		classes: 'always'
 	}],
 	'prefer-template': 'error',
 	'quote-props': ['error', 'as-needed'],
-	quotes: 'off',
-	'@typescript-eslint/quotes': ['error', 'single'],
-	'space-before-function-paren': 'off',
-	'@typescript-eslint/space-before-function-paren': 'error',
-	semi: 'off',
-	'@typescript-eslint/semi': 'error',
+	...rule.override('quotes', '@typescript-eslint/quotes', ['error', 'single']),
+	...rule.override('semi', '@typescript-eslint/semi', 'error'),
+	...rule.override('space-before-function-paren', '@typescript-eslint/space-before-function-paren', 'error'),
 	'sort-imports': ['error', {
 		allowSeparatedGroups: true,
 		ignoreCase: false,
@@ -42,6 +43,8 @@ module.exports = {
 		ignoreMemberSort: false,
 		memberSyntaxSortOrder: ['single', 'multiple', 'all', 'none']
 	}],
-	'space-before-blocks': ['error', 'always'],
+	...rule.override('space-before-blocks', '@typescript-eslint/space-before-blocks', 'error', 'always'),
+	'space-in-parens': ['error', 'never'],
+	'@typescript-eslint/member-delimiter-style': 'error',
 	'@typescript-eslint/type-annotation-spacing': 'error'
 };